server/requirements.txt
flask
flask-cors
openai-whisper
fpdf
torch
reportlab
